aragats-lightning-repo/yzausig/requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.4
numpy==1.16.2
packaging==19.0
Pillow==6.0.0
pip-review==1.0
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
scipy==1.2.1
six==1.12.0
subprocess32==3.5.3
py-getch==1.0.1
tensorflow==1.13.1
